refactor: give the artifact budget tests their own file
The module is close enough to the size cap that the next production change
to it would push the combined file over.
